+++ python/erl_common/storage.py
@@ -34,5 +34,6 @@ def save_numpy_as_eigen_binary(arr: np.ndarray, filename: str):
         arr = arr.reshape(-1, 1)
     assert arr.ndim == 2, "Only support 1D or 2D array"
     with open(filename, "wb") as f:
+        f.write(np.array(arr.size).astype(np.int64).tobytes())
         f.write(np.array(arr.shape).astype(np.int64).tobytes())
         f.write(arr.tobytes("F"))