+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+# 2024-07-25
+
+- Add support to change logging level globally
+- Fix python binding error in Ubuntu 20.04
+
 # 2024-07-24
+
 - Add erl::common::PangolinPlotterImage
 
 # 2024-07-22
+
 - Add Tracy
 - Update setup.py to set `ERL_USE_TRACY=OFF` by default
 - Improve Eigen IO